renamed translation.py -> translator.py
